requirements.txt
pyzmq==16.0.2
japronto==0.1.1
python-dotenv==0.7.1
celery==4.1.0
pytest==3.0.5
redis==2.10.5
requests
flower